refactor: routine 관련 리팩토링 Related to: #2

+++ src/main/java/com/umc5th/muffler/entity/WeeklyRoutineExpense.java
@@ -30,7 +30,7 @@ public class WeeklyRoutineExpense {
     private Integer term;
 
     @Column(nullable = false)
-    private Integer cost;
+    private Long cost;
 
     @Column(nullable = true)
     private LocalDate endDate;
